% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_attempts.R
\name{configure_attempts}
\alias{configure_attempts}
\alias{show_attempts_configuration}
\title{Configure an Attempt for Specific Users}
\usage{
configure_attempts(
  user_ids,
  opens,
  closes,
  max_attempts,
  timelimit,
  grace_period
)

show_attempts_configuration()
}
\arguments{
\item{user_ids}{a character vector of user id's, matched against the user id as determined by the authentication
provider specified in \code{\link[=exam_config]{exam_config()}}.}

\item{opens, closes}{the opening and closing time of the exam for the given users,
in the format \emph{YYYY-MM-DD HH:MM Timezone} (e.g., \verb{2020-10-15 19:15 Europe/Vienna}
for October 15th, 2020 at 19:15 in Vienna, Austria).
The exam is only accessible within this time frame.
Can also be \code{NA} to either make the exam available immediately, indefinitely, or both.}

\item{max_attempts}{maximum number of attempts allowed. Can also be \code{Inf} to allow for unlimited number of attempts.}

\item{timelimit}{the time limit for a single attempt either as a single number in minutes or as \emph{HH:MM}.
Can also be \code{Inf} to give users unlimited time.}

\item{grace_period}{number of seconds of "grace" period given to users before an active attempt is disabled.}
}
\value{
\code{configure_attempts()} returns the updated settings, invisibly.

\code{show_attempts_configuration()} creates an output table to show the attempts configurations.
It only shows the global configuration as well as users which have \emph{different} settings than the global settings.
}
\description{
Override the global settings for attempts for specific users. Any unspecified setting will fall back to
the global setting. This function can only be called from a \emph{server-start} context.

For verifying the attempts configurations, the current configurations and all overrides can be displayed in an
interactive document via \code{show_attempts_configuration()}. This is only meant for verification purposes and should
not be used in the actual exam!
}
