% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise_server.R
\name{evaluate_exercise}
\alias{evaluate_exercise}
\title{Evaluate Exercise}
\usage{
evaluate_exercise(
  user_code,
  support_code,
  chunk_options,
  status_messages,
  envir,
  label
)
}
\arguments{
\item{user_code}{character vector of user code.}

\item{support_code}{list of support codes.}

\item{chunk_options}{chunk options for the \emph{user} code.}

\item{envir}{environment in which to execute the user and support code.}

\item{label}{the exercise label}
}
\description{
Evaluate the user code of an exercise. This compiles an Rmd file with the given code
and calls \code{\link[rmarkdown:render]{rmarkdown::render()}} in the correct environment.
}
\keyword{internal}
