% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_format.R
\name{exam_document}
\alias{exam_document}
\title{Exam Output Format}
\usage{
exam_document(
  id = "exam",
  version = "0.1",
  use_cdn = FALSE,
  sections = list(),
  question_context = "default",
  points_context = "info",
  exercise_button_context = points_context,
  ...
)
}
\arguments{
\item{id}{the exam id string}

\item{version}{the exam version string}

\item{use_cdn}{load javascript libraries from external content delivery networks (CDNs).
Use this if the shiny server is slow at serving static resources, but beware of the downsides of
relying on content from a third-party!}

\item{sections}{section configuration, see \code{\link[=sections_options]{sections_options()}} for details.}

\item{question_context, points_context, exercise_button_context}{contextual style classes for
question panels, points labels, exercise buttons, and section navigation buttons.
Can be \code{default} or any of the bootstrap 3 contextual classes listed at
\url{https://getbootstrap.com/docs/3.3/css/#helper-classes-colors}.
These can also be set on a per-question or per-exercise level by specifying code chunk options
\code{exam.question_context}, \code{exam.points_context} and \code{exercise.button_context}.
A section-specific button context can be set via \code{\link[=section_specific_options]{section_specific_options()}}.}

\item{...}{passed on to \link[rmarkdown:html_document]{html_document()}. Parameters \code{section_divs} and \code{toc} are not
supported.

These status messages can also be overwritten by setting the exercise chunk options \code{exercise.status_xyz}, where
\code{xyz} is one of the status message types listed above.}
}
\description{
Exam Output Format
}
