% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise_options.R
\name{exercise_options}
\alias{exercise_options}
\title{Exercise Options}
\usage{
exercise_options(
  title,
  button,
  timelimit,
  lines,
  autocomplete,
  df_print,
  points,
  setup,
  solution,
  checker,
  label
)
}
\arguments{
\item{title}{title of the exercise. Default: \emph{exercise.panelTitle} from \code{\link[=status_messages]{status_messages()}}.}

\item{button}{label of the run exercise button. Default: \emph{exercise.buttonLabel} from \code{\link[=status_messages]{status_messages()}}.}

\item{timelimit}{time limit for the code to run (in seconds). Default: 5.}

\item{lines}{the minimum number of lines in the code editor. Default: 5.}

\item{autocomplete}{enable or disable auto-completion in the code editor. Note that this sends many requests
to the server and could pose performance issues if many users access the exam at the same time. Default: \code{FALSE}.}

\item{df_print}{method used for printing data frames created by the user code. Default to use the same
as the R-markdown document, but can be changed if needed. See argument \code{df_print} in
\link[rmarkdown:html_document]{html_document()} for details.}

\item{points}{the default number of points an exercise is worth. Default: 1.}

\item{setup}{the name of an R code chunk which is evaluated before the user code.}

\item{solution}{the name of an R code chunk which gives the solution to the exercise code chunk.
This really only makes sense if set on a per-chunk basis.}

\item{checker}{a function (or the name of a function) which checks the R code in the exercise chunk.}

\item{label}{a label to help screen readers describe the purpose of the input element.
Default: \emph{exercise.label} from \code{\link[=status_messages]{status_messages()}}.}
}
\description{
Global options for exercise chunks. Each of these options can be overwritten in the individual exercise chunks by
specifying the \code{exercise.xyz} chunk option (\code{xyz} is one of the parameter names below).
}
\seealso{
Other exercise configuration: 
\code{\link{exercise_evaluator}}
}
\concept{exercise configuration}
