% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_exercise_chunk.R
\name{exercise_result}
\alias{exercise_result}
\alias{exercise_result_error}
\alias{exercise_result_empty}
\alias{exercise_result_timeout}
\title{Create Exercise Results}
\usage{
exercise_result(
  feedback,
  html_output,
  error_message,
  timeout_exceeded = FALSE,
  severity,
  label = NULL
)

exercise_result_error(
  error_message,
  severity = c("error", "warning", "info"),
  timeout_exceeded = FALSE,
  label = NULL
)

exercise_result_empty(label = NULL, severity)

exercise_result_timeout(label = NULL)
}
\arguments{
\item{feedback}{a message to be displayed underneath the code chunk.}

\item{html_output}{the HTML code to be inserted into the exercise's output container.}

\item{error_message}{if present, \code{severity} is set to \code{error} and the error message is used as feedback.}

\item{timeout_exceeded}{if \code{TRUE}, \code{severity} is set to \code{error} and the timeout error message is shown.}

\item{severity}{can be one of \code{error}, \code{warning}, or \code{info}.}

\item{label}{the label of the exercise chunk. If missing, globally defined status messages are displayed
instead of the chunk-specific messages.}
}
\description{
Create error results from a custom exercise evaluator.

\code{exercise_result_error()} is the preferred way to create an error result.

\code{exercise_result_empty()} creates an exercise result representing an "empty" output.

\code{exercise_result_timeout()} is a short-hand version for the common problem of the R code timing out.
}
\details{
The function \code{exercise_result_timeout()} uses the correct message configured for the exam (or code chunk).
}
