% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_sections.R, R/question_types.R
\name{knit_print.examinr_section_chunk}
\alias{knit_print.examinr_section_chunk}
\alias{knit_print.examinr_section_end}
\alias{knit_print.examinr_question}
\title{Overrides for knit_print.}
\usage{
\method{knit_print}{examinr_section_chunk}(x, ...)

\method{knit_print}{examinr_section_end}(x, ...)

\method{knit_print}{examinr_question}(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\description{
Overrides for knit_print.

Overrides for knit_print.

Overrides for knit_print.
}
\keyword{internal}
