% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_provider.R
\name{password_auth}
\alias{password_auth}
\alias{ui_token_auth}
\title{Authenticate Users Via User Interface}
\usage{
password_auth(
  users,
  username_label = "Username",
  password_label = "Password",
  title = "Login",
  button_label = "Login",
  username_empty = "Username cannot be empty.",
  password_empty = "Passwort cannot be empty.",
  unauthorized = "The username/password are incorrect."
)

ui_token_auth(
  users,
  token_label = "Access token",
  title = "Login",
  button_label = "Login",
  token_empty = "Access token cannot be empty.",
  unauthorized = "The token is incorrect."
)
}
\arguments{
\item{users}{a data frame with user authentication information. See below for details.}

\item{username_label}{label for the username input field.}

\item{password_label}{label for the password input field.}

\item{title}{title of the dialog.}

\item{button_label}{label for the login button.}

\item{username_empty}{error messages for empty username field.}

\item{password_empty}{error messages for empty password field.}

\item{unauthorized}{error message for incorrect username/password.}

\item{token_label}{label for the token input field.}

\item{token_empty}{error messages for empty token field.}
}
\description{
Present the user with a login interface before the exam begins.
\code{password_auth()} requests the user to enter a username and password.

\code{ui_token_auth()} requires the user to enter only an access token.
}
\details{
The \code{password_auth()} authentication provider
matches the given username and password against columns \emph{username} and \emph{password} in data frame \code{users}.
The \emph{password} column must contain \strong{hashed} passwords created with \code{\link[sodium:password]{sodium::password_store()}}
(e.g., \code{sodium::password_store("password")}, or a SHA-256 HMAC created with \code{\link[digest:hmac]{digest::hmac()}}
(\code{digest::hmac("exam_id", "password", algo = "sha256")}).

The \code{ui_token_auth()} authentication provider
matches the given token and against column \emph{token} in data frame \code{users}.
The token can be any character data.

For both authentication providers, if the \code{users} data frame also contains columns \emph{display_name}
and/or \emph{grading}, they are added to the user object
(and used as described in the \link[=authentication_provider]{authentication provider documentation}).
}
\seealso{
Other access configuration: 
\code{\link{authentication_provider}},
\code{\link{configure_attempts}()},
\code{\link{rsconnect_auth}()}
}
\concept{access configuration}
