% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_provider.R
\name{rsconnect_auth}
\alias{rsconnect_auth}
\title{RStudio Connect Authentication Provider}
\usage{
rsconnect_auth(grading_groups = NULL)
}
\arguments{
\item{grading_groups}{users who are members of any of these groups will have grading permissions.}
}
\value{
an authentication provider to use in \code{\link[=exam_config]{exam_config()}}.
}
\description{
Simple \link[=authentication_provider]{authentication provider} using the current user authenticated with RStudio Connect.
The \code{user_id} is the user's username and grading permission are determined by the user's group membership.
}
\examples{
# Create an authentication provider which grants grading permissions to users which
# are members of either the "instructor" or the "assistant" group.
auth_prov <- rsconnect_auth(c('instructor', 'assistant'))
exam_config(auth_provider = auth_prov)

}
\seealso{
Other access configuration: 
\code{\link{authentication_provider}},
\code{\link{configure_attempts}()},
\code{\link{ui_auth}()}
}
\concept{access configuration}
