% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_options.R
\name{sections_options}
\alias{sections_options}
\alias{section_specific_options}
\title{Exam Sections Options}
\usage{
sections_options(render, order, fixed, next_button_label, .force = TRUE)

section_specific_options(
  section,
  next_button_label,
  next_button_context,
  fix_order
)
}
\arguments{
\item{render}{render the sections on the server (\code{"server"}) or the client (\code{"client"}).
If rendered on the server, the section content will only be sent to the client (i.e., the browser)
when the section is displayed. This ensures users do not see the section content before navigating to the section.}

\item{order}{should be section order be fixed (\code{"fixed"}) or randomized (\code{"random"}) for each user.}

\item{fixed}{if \code{order="random"}, a character vector of sections (their names) which will be fixed the order.}

\item{next_button_label}{the label for the button at the end of a section. If \code{NA}, no button will be displayed.
Default button label is \emph{Next section}.}

\item{.force}{force the given options over already specified options. Default to \code{TRUE}.}

\item{section}{name of the section for which to set specific options.}

\item{next_button_context}{bootstrap 3 contextual class applied to the button. Takes precedence over the global
specification in the \code{\link[=exam_document]{exam_document()}} output format. See
\url{https://getbootstrap.com/docs/3.3/css/#buttons-options} for a list of available classes.}

\item{fix_order}{logical if the section should remain fixed in order.}
}
\description{
Control the behavior of the individual \emph{sections} of an exam.
Each exam contains one or more sections, which are shown progressively to the user.

With \code{section_specific_options()}, options can be specified for individual sections identified via a section id.
}
\details{
The sections options can be set in the metadata of the exam file and/or through the \code{sections_options()} function.
Options set via \code{sections_options()} and \code{section_specific_options()} take precedence over options set in
the metadata, if \code{.force=TRUE} (as by default).

Sections are identified by the name.
For example, a section with title \emph{# Section 1: Multiple Choice Questions} would be identified by
\code{"Section 1: Multiple Choice Questions"}.
Note that pandoc heading identifiers (anything in \code{{}} at the end of the section title are omitted from the name.
}
