% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_utilities.R
\name{setup_database}
\alias{setup_database}
\title{Setup a DBI-Backed Database for Exams}
\usage{
setup_database(db_con)
}
\description{
Helper function to setup a "default" database in a form suitable for use with
\code{\link[=dbi_storage_provider]{dbi_storage_provider()}}.
This function creates two tables (\emph{attempts} and \emph{sections}) and the proper indices
to ensure uniqueness of entries.
}
\details{
Currently implemented only for RPostgres or RSQLite backends.
}
\seealso{
Other storage configuration: 
\code{\link{dbi_storage_provider}()},
\code{\link{storage_provider}}
}
\concept{storage configuration}
