% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_messages.R
\name{status_message_template}
\alias{status_message_template}
\alias{status_messages}
\title{Customize Status Messages}
\usage{
status_message_template(file)

status_messages(file, messages)
}
\arguments{
\item{file}{path to the messages file. For \code{status_message_template()}, this is where the template is saved
(if \code{NULL}, only returns the default messages). For \code{status_messages()}, the file to read the status messages
from.}

\item{messages}{Optionally specify the messages via an R list, in the same format as returned by
\code{status_message_template()}. If \code{messages} is set, \code{file} is ignored with a warning.}
}
\value{
\code{status_message_template()} invisibly returns the default messages as R list.
\code{status_messages()} invisibly returns the new status messages as R list.
}
\description{
Customize the status messages displayed at various times and states during an exam.
A template with the default status messages can be saved as a YAML text file with \code{status_message_template()}.
}
\details{
Note that \code{status_message_template()} should not be called from within an exam document. Rather, generate the
template once, modify as necessary, and use in the exam document by specifying the file name in the \code{\link[=exam_document]{exam_document()}}
or by calling \code{use_status_messages()}
}
