% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_provider.R
\name{authentication_provider}
\alias{authentication_provider}
\title{Authentication Providers}
\arguments{
\item{session}{the shiny session object.}
}
\value{
a list with user information. The list \strong{must} contain an element \code{user_id} with the user's unique
identifier as character (it will be forcefully cast to \code{character}).
The user object may also contain the following
\describe{
\item{\code{display_name}}{a character string which will be displayed for grading. If missing, the user id is
displayed for grading.}
\item{\code{grading}}{if \code{TRUE}, the user has permissions to grade attempts.}
}
The user object can optionally include any number of additional entries which will be stored alongside
the exam data.
\strong{Warning:} this could pose serious privacy issues. Either don't add additional information or ensure the data is
stored encrypted and secure.

Any additional information will be proxied to all callback functions receiving a \code{user} object.
}
\description{
Every exam attempt is tied to a single \emph{user}.
The exam document format does not determine the user, but invokes the authentication provider configured with
\code{\link[=exam_config]{exam_config()}} to determine the user's identification.
}
\examples{
# Use a "dummy" authentication provider which assigns everyone the same user id
# and does not allow anyone to grade submitted attempts.
exam_config(auth_provider = function (session) {
  return(list(
    user_id = "E-L",
    display_name = "Eager Learner",
    grading = FALSE
  ))
})

}
\seealso{
\code{\link[=rsconnect_auth]{rsconnect_auth()}} for an authentication provider using the user information from RStudio Connect.

\code{\link[=password_auth]{password_auth()}} and \code{\link[=ui_token_auth]{ui_token_auth()}} for authentication providers displaying a login user interface
before the exam begins.

Other access configuration: 
\code{\link{configure_attempts}()},
\code{\link{password_auth}()},
\code{\link{rsconnect_auth}()}
}
\concept{access configuration}
