% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_provider.R
\name{authentication_provider}
\alias{authentication_provider}
\alias{rsconnect_authentication_provider}
\title{Authentication Providers}
\usage{
rsconnect_authentication_provider(shiny_session)
}
\arguments{
\item{shiny_session}{the shiny session object.}
}
\value{
a list with user information. The list \strong{must} contain an element \code{user_id} with the user's unique
identifier as character (it will be forcefully cast to \code{character}).
The user information list can optionally include any number of additional objects, which will be stored alongside
the exam data.
\strong{Warning:} this could pose serious privacy issues. Either not add additional information or ensure the data is
stored encrypted and secure.

Any additional information will be proxied to all callback functions receiving a \code{user} object.
}
\description{
Every exam attempt is tied to a single \emph{user}.
Each attempt can get a different randomization of the exam and all attempts are tied to a user.
The exam document format does not determine the user, but invokes the authentication provider configured with
\code{\link[=exam_config]{exam_config()}} to determine the user's identification.
}
\section{Functions}{
\itemize{
\item \code{rsconnect_authentication_provider}: use the name of the user authenticated with RStudio Connect as user id.
}}

