% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_provider.R
\name{dbi_storage_provider}
\alias{dbi_storage_provider}
\title{DBI-Backed Exam Storage}
\usage{
dbi_storage_provider(conn, table_name, hash_user = FALSE)
}
\arguments{
\item{conn}{a \link[DBI:DBIConnection-class]{DBIConnection} object, as returned by \link[DBI:dbConnect]{dbConnect()}}

\item{table_name}{the name of the table to store the exam data in.}

\item{hash_user}{store the user id as a SHA256 hash to protect the identity.}
}
\description{
Store exam data in an SQL database using DBI.
\strong{Note that exam data may contain sensitive information.}
To avoid potential privacy issues, the DBI storage provider supports hashing the user id, but this may require
external identification of users for grading.
}
\details{
Exam data is stored in a table which requires the following columns:
\itemize{
\item \code{user_id} stored as character data. The required length depends on the authentication provider and if hashing is
used.
\item \code{exam_id} stored as character data. Ensure it can hold enough characters for the exam ids used in your exams.
\item \code{exam_version} stored as character data. Ensure it can hold enough characters for the exam version strings
used in your exams.
\item \code{timestamp} the UTC time at which the section data was last saved (of type \emph{TIMESTAMP})
\item \code{section_data} stored as text of arbitrary length.
}

A sample definition of the table for PostgreSQL is as follows:\preformatted{CREATE TABLE section_data (
  id serial PRIMARY KEY,
  user_id character varying(64) NOT NULL,
  exam_id character varying(255) NOT NULL,
  exam_version character varying(64) NOT NULL,
  timestamp timestamp NOT NULL DEFAULT current_timestamp,
  section_data text);
CREATE INDEX section_data_index on section_data (user_id, exam_id, exam_version);
}
}
