% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise_evaluators.R
\name{exercise_evaluator}
\alias{exercise_evaluator}
\alias{local_evaluator}
\alias{psock_evaluator}
\title{Exercise Evaluator}
\usage{
local_evaluator(expr, envir, label, timelimit, ...)

psock_evaluator(expr, envir, label, timelimit, ...)
}
\arguments{
\item{expr}{an (un-evaluated) expression to be evaluated.}

\item{envir}{the environment in which \code{expr} is to be evaluated.}

\item{label}{the label of the exercise chunk.}

\item{timelimit}{the configured time limit in seconds.}

\item{...}{additional parameters for future extensions.}
}
\value{
a list of functions, namely
\itemize{
\item \code{completed()} returns \code{TRUE} if the evaluator is finished, \code{FALSE} otherwise.
\item \code{kill()} kill/interrupt the evaluator. This will be called when the time limit has passed.
\item \code{result()} return the result from evaluating \code{expr}.
\item \code{start()} start the evaluation of \code{expr}.
}
}
\description{
Configure how exercises are evaluated.
An exercise evaluator is a function taking the arguments described below and returning a list of functions.
}
\section{Functions}{
\itemize{
\item \code{local_evaluator}: evaluate the user code in the current R process.

\item \code{psock_evaluator}: evaluate the user code in a separate R process, created with
\code{\link[parallel:makeCluster]{parallel::makePSOCKcluster()}}.
}}

