% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_options.R
\name{exercise_options}
\alias{exercise_options}
\title{Exercise Options}
\usage{
exercise_options(
  title,
  button,
  button_context,
  timelimit,
  lines,
  autocomplete,
  df_print,
  points,
  status_notrun,
  status_invalid,
  status_error,
  status_success,
  status_empty,
  status_timeout
)
}
\arguments{
\item{title}{title of the exercise. Default: \emph{Coding question.}}

\item{button}{label of the run exercise button. Default: \emph{Run code.}}

\item{button_context}{the bootstrap contextual class applied to the button. Default: \emph{default}. See
\url{https://getbootstrap.com/docs/3.3/css/#buttons-options} for a list of supported context classes.}

\item{timelimit}{time limit for the code to run (in seconds). Default: 5.}

\item{lines}{the minimum number of lines in the code editor. Default: 5.}

\item{autocomplete}{enable or disable auto-completion in the code editor. Default: \code{FALSE}.}

\item{df_print}{method used for printing data frames created by the user code. Default to use the same
as the R-markdown document, but can be changed if needed. See argument \code{df_print} in
\link[rmarkdown:html_document]{html_document()} for details.}

\item{points}{the default number of points an exercise is worth. Default: 1.}

\item{status_notrun}{status message if the exercise code has not yet been executed in this session. Default message:
\emph{Run the code by clicking the button in the top right corner of the exercise.}}

\item{status_invalid}{status message if the exercise code is invalid (currently only checked for R code). The
message is followed by detailed information about the problem.
Default message: _R code is invalid: _}

\item{status_error}{status message if the exercise code has resulted in an unknown error. Default message:
\emph{Cannot run the code.}}

\item{status_success}{status message if the exercise code has been run successfully and the output is shown below.
Default message: \emph{Code executed successfully.}}

\item{status_empty}{status message if the exercise code has been run successfully, but did not produce any output.
Default message: \emph{Code executed successfully, but no output generated.}}

\item{status_timeout}{status message if the exercise code takes too long to run. Default message:
\emph{Code takes too long to execute.}}
}
\description{
Global options for exercise chunks. Each of these options can be overwritten in the individual exercise chunks by
specifying the \code{exercise.xyz} chunk option (\code{xyz} is one of the parameter names below).
}
