% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise_options.R
\name{exercise_options}
\alias{exercise_options}
\title{Exercise Options}
\usage{
exercise_options(
  title,
  button,
  button_context,
  timelimit,
  lines,
  autocomplete,
  df_print,
  points,
  setup,
  solution,
  checker,
  label
)
}
\arguments{
\item{title}{title of the exercise. Default: \emph{Coding question.}}

\item{button}{label of the run exercise button. Default: \emph{Run code.}}

\item{button_context}{the bootstrap contextual class applied to the button. Default: \emph{default}. See
\url{https://getbootstrap.com/docs/3.3/css/#buttons-options} for a list of supported context classes.}

\item{timelimit}{time limit for the code to run (in seconds). Default: 5.}

\item{lines}{the minimum number of lines in the code editor. Default: 5.}

\item{autocomplete}{enable or disable auto-completion in the code editor. Default: \code{FALSE}.}

\item{df_print}{method used for printing data frames created by the user code. Default to use the same
as the R-markdown document, but can be changed if needed. See argument \code{df_print} in
\link[rmarkdown:html_document]{html_document()} for details.}

\item{points}{the default number of points an exercise is worth. Default: 1.}

\item{setup}{the name of an R code chunk which is evaluated before the user code.}

\item{solution}{the name of an R code chunk which gives the solution to the exerise code chunk.
This really only makes sense if set on a per-chunk basis.}

\item{checker}{a function (or the name of a function) which checks the R code in the exercise chunk.}

\item{label}{a label to help screen readers describe the purpose of the input element.}
}
\description{
Global options for exercise chunks. Each of these options can be overwritten in the individual exercise chunks by
specifying the \code{exercise.xyz} chunk option (\code{xyz} is one of the parameter names below).
}
