% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_attempts.R, R/exam_status_ui.R,
%   R/question_types_ui.R, R/section_ui.R
\name{knit_print.examinr_attempts_config_display}
\alias{knit_print.examinr_attempts_config_display}
\alias{knit_print.examinr_exam_status}
\alias{knit_print.examinr_question}
\alias{knit_print.examinr_section_init}
\alias{knit_print.examinr_section_start}
\alias{knit_print.examinr_section_chunk}
\alias{knit_print.examinr_section_end}
\title{Overrides for knit_print.}
\usage{
\method{knit_print}{examinr_attempts_config_display}(x, ...)

\method{knit_print}{examinr_exam_status}(x, ...)

\method{knit_print}{examinr_question}(x, ...)

\method{knit_print}{examinr_section_init}(x, ...)

\method{knit_print}{examinr_section_start}(x, ...)

\method{knit_print}{examinr_section_chunk}(x, ...)

\method{knit_print}{examinr_section_end}(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\description{
Overrides for knit_print.

Overrides for knit_print.

Overrides for knit_print.

Overrides for knit_print.

Overrides for knit_print.

Overrides for knit_print.

Overrides for knit_print.
}
\keyword{internal}
