% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_types_ui.R
\name{mc_question}
\alias{mc_question}
\alias{answer}
\title{Create a Multiple-Choice Question}
\usage{
mc_question(
  title,
  ...,
  points = 1,
  nr_answers = 5,
  random_answer_order = TRUE,
  checkbox = TRUE,
  label = "Select the correct answer(s).",
  hide_label = FALSE,
  min_points = 0,
  mandatory = FALSE,
  id = NULL,
  title_container = h6,
  static_title = NULL
)

answer(
  label,
  correct = FALSE,
  always_show = FALSE,
  weight = c(0, 1),
  correct_quoted = FALSE
)
}
\arguments{
\item{title}{question title. Markdown is supported and by default the title will be rendered dynamically if it
contains inline R code.}

\item{...}{answer options created with \code{\link[=answer]{answer()}}.}

\item{points}{total number of points for the question. Set to \code{NULL} to not give any points for the question.
Will be shown next to the question text, formatted using the \code{points_format} given in \code{\link[=exam_config]{exam_config()}}.}

\item{nr_answers}{maximum number of answers to display. At least one correct answer will always be shown.
If \code{NULL}, all answer options are shown. If a vector of length two, the first number specifies the number of
\emph{correct} answers to be shown and the second number specifies the number of \emph{incorrect} answers to be shown.}

\item{random_answer_order}{should the order of answers be randomized? Randomization is unique for every user.}

\item{checkbox}{make a multiple response question, i.e., allow the user to select more than one answer.}

\item{label}{answer text to show. Supports markdown formatting and inline code chunks of the form \code{`r value`}.}

\item{hide_label}{hide the label from non-screen readers.}

\item{min_points}{if \code{points} multiplied by the the sum of the weights of all selected answer options is negative,
where to cut off negative points. If \code{NULL}, there is no lower bound.}

\item{mandatory}{is this question mandatory for submitting the section? If \code{TRUE}, a user can only navigate to the
next section if the question is answered.}

\item{id}{question identifier to be used instead of the code chunk label. Must only contain
characters \code{a-z}, \code{A-Z}, \code{0-9}, dash (\code{-}), or underscore (\verb{_}).}

\item{title_container}{a function to generate an HTML element to contain the question title.}

\item{static_title}{if \code{NULL}, the title will be rendered statically if it doesn't contain inline R code, otherwise
it will be rendered dynamically. If \code{TRUE}, always render the title statically. If \code{FALSE}, always render the
title dynamically on the server.}

\item{correct}{an expression which, evaluated in the rendering environment, determines if the answer is correct
or not. Can also be a simple logical.}

\item{always_show}{even if answer options are randomized, always show this answer.}

\item{weight}{the weight of this answer, if selected. By default, correct answers have a weight of 1, and incorrect
answers have a weight of 0. See details for more information.
A negative weight will subtract that proportion of available points from the user's
points (e.g., to penalize for selecting incorrect answer options).
The sum of all displayed \strong{positive} weights is standardized to sum to 1,
\strong{but negative weights are not standardized.}}

\item{correct_quoted}{is the expression given in \code{correct} quoted or not?}
}
\description{
Show a multiple-choice or single-choice question.
}
\details{
The \code{weight} of an answer option determines how many points are awarded/subtracted for selecting this answer option.
If \code{correct} is an expression and correctness of an answer option can only be determined at run-time, \code{weight}
can be a numeric vector with two elements. The first element is the weight applied if the answer is wrong, the
second element is the weight applied if the answer is correct.

The sum of weights of all displayed correct answers is standardized to sum to 1 (i.e., users selecting
all displayed correct answers will get all points), but negative weights are used as given.
For example, if the weight of an incorrect answer option is set to -0.5 and the total number of points available
for a question is 4, selecting this answer option will reduce the number of points awarded by 2.
}
\seealso{
Other exam question types: 
\code{\link{text_question}()}
}
\concept{exam question types}
