% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_config.R
\name{section_config}
\alias{section_config}
\title{Exam Section Configuration}
\usage{
section_config(section, next_button_label, next_button_context, fix_order)
}
\arguments{
\item{section}{name of the section for which to set configuration options.}

\item{next_button_label}{the label for the button at the end of a section.
Default button label is \emph{Submit answers}.
If \code{progressive=FALSE}, only the button in the last section will be displayed.
Buttons can be hidden for specific sections via \code{section_specific_options(next_button_label=NA)}. This is only
useful for progressive exams where the last section is only for information purposes.}

\item{next_button_context}{bootstrap 3 contextual class applied to the button. See
\url{https://getbootstrap.com/docs/3.3/css/#buttons-options} for a list of available classes.}

\item{fix_order}{logical if the section should remain fixed in order.}
}
\description{
Control the behavior of each individual section of an exam.
Each exam contains one or more sections, which are shown either all-at-once or progressively to the user.
}
\details{
The section configuration can be set in the metadata of the exam file and with the \code{section_config()} function.
Options set via \code{section_config()} take precedence over options set in the metadata.

The last section in the exam document is assumed to be a "final remark" and displayed to the user after
an attempt is finished, without possibility of submitting the section.

Sections are identified by their name.
For example, a section with title \emph{# Section 1: Multiple Choice Questions} would be identified by
\code{"Section 1: Multiple Choice Questions"}.
Note that pandoc heading identifiers (anything in \code{{}} at the end of the section title) are omitted from the name.
}
