% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_provider.R
\name{storage_provider}
\alias{storage_provider}
\title{Storage Providers}
\value{
a list of callback functions:
\itemize{
\item \code{save_section_data(section_data, ...)}: Save the given section data.
The section data is an R list with with arbitrary elements. Should return \code{TRUE} if the data is saved
successfully or \code{FALSE} in case of an error.
\item \code{get_section_data(...)}: Return the most recent section data for the given user.
The returned object must be in the same format as the section data provided to \code{save_section_data()}.
In case of an error, should return \code{NULL}.
}

See the section \emph{Callback functions} for more details on the additional arguments \code{...}.
}
\description{
Exam data is stored via \emph{storage providers}.
Each user action will trigger a storage event for which a callback function is called.
See \code{\link[=dbi_storage_provider]{dbi_storage_provider()}} for an implementation storing/retrieving section data from a DBI backend.
}
\section{Callback functions}{

All callback functions must accept the following arguments:
\itemize{
\item \code{user} the user object (as returned from the configured \link[=authentication_provider]{authentication provider}).
\item \code{exam_id} the exam id as given in the Rmd file.
\item \code{exam_version} the exam version string as given in the Rmd file.
\item \code{...} for future extensions.
}

Furthermore, each callback function is called with the following optional arguments:
\itemize{
\item \code{shiny_session} the shiny session object.
}
}

