% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_types_ui.R
\name{text_question}
\alias{text_question}
\title{Create a Question With Text Input}
\usage{
text_question(
  title,
  points = 1,
  type = c("textarea", "text", "numeric"),
  width = "100\%",
  height = NULL,
  placeholder = NULL,
  title_container = h6,
  static_title = NULL,
  accuracy = 0.001,
  solution = NULL,
  solution_quoted = FALSE,
  label = "Type your answer below.",
  hide_label = FALSE
)
}
\arguments{
\item{title}{question title. Markdown is supported and by default the title will be rendered dynamically if it
contains inline R code.}

\item{points}{total number of points for the question. Set to \code{NULL} to not give any points for the question.
Will be shown next to the question text, formatted using the \code{points_format} given in \code{\link[=exam_config]{exam_config()}}.}

\item{type}{can be one of \code{textarea} (default), \code{text} or \code{numeric}.}

\item{width}{the width of the input, e.g., '300px' or '100\%'; see \code{\link[shiny:reexports]{shiny::validateCssUnit()}}.}

\item{height}{the height of the input, e.g., '300px' or '100\%'; see \code{\link[shiny:reexports]{shiny::validateCssUnit()}}. Only relevant for
\code{type="textarea"}.}

\item{placeholder}{A character string giving the user a hint as to what can be entered into the control.
Internet Explorer 8 and 9 do not support this option.}

\item{title_container}{a function to generate an HTML element to contain the question title.}

\item{static_title}{if \code{NULL}, the title will be rendered statically if it doesn't contain inline R code, otherwise
it will be rendered dynamically. If \code{TRUE}, always render the title statically. If \code{FALSE}, always render the
title dynamically on the server.}

\item{solution}{an expression to compute the solution to the answer.
The expression is evaluated in the environment returned by the data provider set up via \code{\link[=exam_config]{exam_config()}}.
See below for details on how to auto-grade questions of type \emph{numeric}.}

\item{solution_quoted}{is the \code{solution} expression quoted?}

\item{label}{a label to help screen readers describe the purpose of the input element.}

\item{hide_label}{hide the label from non-screen readers.}

\item{markdown_html()}{numeric accuracy for grading inputs with \code{type="numeric"}.}
}
\description{
Render a question which asks the user to input a value. Questions of \emph{numeric} support auto-grading.
}
\details{
The solution is computed by evaluating the expression given in \code{solution} in the rendering environment.
The expression must yield either a single character string or, if the question is of type \emph{numeric}, a numeric value.
The result of the expression is rendered with \link[commonmark:commonmark]{commonmark}.
To auto-grade \emph{numeric} questions, the \code{solution} expression can yield a numeric value or a character value with
attribute \code{answer}. This number is compared against the user's answer (with tolerance \code{accuracy}).
}
