% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_types.R
\name{text_question}
\alias{text_question}
\title{Create a Text Question}
\usage{
text_question(
  title,
  points = 1,
  type = c("textarea", "text", "numeric"),
  width = "100\%",
  height = NULL,
  placeholder = NULL,
  title_container = h5,
  static_title = NULL,
  min = NA,
  max = NA,
  step = NA,
  accuracy = step,
  label = NULL
)
}
\arguments{
\item{title}{question title. Markdown is supported and by default the title will be rendered dynamically if it
contains inline R code.}

\item{points}{total number of points for the question. Set to \code{NULL} to not give any points for the question.
Will be shown next to the question text, formatted using the \code{points_format} given in \code{\link[=exam_options]{exam_options()}}.}

\item{type}{can be one of \code{textarea} (default), \code{text} or \code{numeric}.}

\item{width}{the width of the input, e.g., '300px' or '100\%'; see \code{\link[shiny:reexports]{shiny::validateCssUnit()}}.}

\item{height}{the height of the input, e.g., '300px' or '100\%'; see \code{\link[shiny:reexports]{shiny::validateCssUnit()}}. Only relevant for
\code{type="textarea"}.}

\item{placeholder}{A character string giving the user a hint as to what can be entered into the control.
Internet Explorer 8 and 9 do not support this option.}

\item{title_container}{a function to generate an HTML element to contain the question title.}

\item{static_title}{if \code{NULL}, the title will be rendered statically if it doesn't contain inline R code, otherwise
it will be rendered dynamically. If \code{TRUE}, always render the title statically. If \code{FALSE}, always render the
title dynamically on the server.}

\item{min, max, step}{if \code{type="numeric"}, gives the minimum and maximum allowed value as well as the interval
for stepping between min and max.}

\item{accuracy}{numeric accuracy for grading inputs with \code{type="numeric"}.}

\item{label}{an optional label displayed above the input element.}
}
\description{
Text questions are not auto-graded.
}
